refactor: move keymap, otp, autocommand, to core and move all plugin  to plugin folder
