+++ lua/plugins/lint copy.lua
