+++ lua/plugins/todo-comments.lua
@@ -1,0 +1,1 @@
+return { 'folke/todo-comments.nvim', event = 'VimEnter', dependencies = { 'nvim-lua/plenary.nvim' }, opts = { signs = false } }
