+++ lua/plugins/vim-sleuth.lua
@@ -1,0 +1,3 @@
+return { -- Detect tabstop and shiftwidth automatically
+  'tpope/vim-sleuth',
+}
